Don't strip off the second byte of the octal mode.
The stat here pulls the raw hex mode of the object being statted. While
it is arguable we can skip honoring the first octal digit (which
indicates whether it is a file, directory, symlink, etc), the next one
(suid, sgid, sticky) is important and should not be stripped.  This
commit un-strips the mode.