Unify file handling for local and remote transports (#189)
Local and remote transports expose different file interfaces. This
is because local transport uses LinuxFile with some methods
overriden with Ruby's File calls whereas remote transport uses
the appropriate file classes from Train::Extras. This causes
problems when users are running locally on Windows because some
LinuxFile implementations are not appropriate there.

There are two potential compat breaks with this change:
1. pw_username is no longer provided for local transport
2. pw_groupname is no longer provided for local transport

These are not available for remote transport, so this doesn't
seem like it should be a problem for clients.

Signed-off-by: Ted Wang <ted.wang@ni.com>