Fix inconsistent link_path behavior (#194)
When calling `path` on a file using the local transport, it would call
Train::Extras::LinuxFile.path which would in turn call read_target_path
if the file was a symlink (which calls the `readlink` OS program). The
return value would be stored in `@link_path` which is then returned whenever
`link_path` is called on the file.

`readlink` does not operate the same way on macOS like it does on Linux
variants. Specifically, in the case of cascading symlinks (a -> b -> c), it will
return "c" on Linux but "b" on macOS. This leads to inconsistent behavior
depending on if you call `path` before `link_path` since `path` stores a value
in `@link_path`. For example:

Example 1 (calling path before link_path):
```
inspec> a = file('/tmp/top_link')
=> File /tmp/top_link
inspec> a.path
=> "/tmp/middle_link"
inspec> a.link_path
=> "/tmp/middle_link"
```

Example 2 (calling link_path before path):
```
inspec> a = file('/tmp/top_link')
=> File /tmp/top_link
inspec> a.link_path
=> "/private/tmp/actual_file"
inspec> a.path
=> "/private/tmp/actual_file"
```

Plus, we have the Ruby stdlib calls we can and should use when
local so Windows, Linux, and macOS users have a consistent experience. Therefore,
this change implements a `path` method override on the Local File class to
avoid LinuxFile from getting in the way.

Signed-off-by: Adam Leff <adam@leff.co>