[win-perf][proof-of-concept] run PS in a server locally
we incur a significant overhead when running windows commands
locally because each command needs to execute within a new
powershell instance. this adds an overhead of approximately
250ms per powershell request.

the changes in the sd/win-perf branch of inspec go a ways
towards reducing the number of powershell commands that are
executed, but they will never overcome the overhead that's
baked in for each one of those calls.

this change is a PROOF OF CONCEPT for what it would look
like to have a long-running powershell process that inspec/train
can send commands to. in this POC, it's implemented as an HTTP
server written in powershell that executed the body of POST
requests and has a special `exit` command to kill the process.

on `master` of inspec, this reduces the execution time of the
`windows-baseline` profile from 30s to 13s. when combined with
the `sd/win-perf` branch of inspec, the execution time is
further reduced to 9s.

Signed-off-by: Stephen Delano <stephen@chef.io>