Fix regressions in 0.29.1 (#219)
* Add unix_mode_mask method to Train::File::Local::Unix (#215)

* Add unix_mode_mask method to Train::File::Local::Unix

During the refactor, the `#unix_mode_mask` method for Unix files was
only added to the Remote class. It's still needed for the Local class.

Signed-off-by: Adam Leff <adam@leff.co>

* Use the sanitized file path for remote linux files (#218)

Had a test start failing recently and tracked it down to remote linux file
resources no longer handling spaces in paths as of the refactor in 0.29.0.

The test passes again after applying this change.

Signed-off-by: Jonathan Hartman <j@hartman.io>

* Adding branches for testing

Signed-off-by: Adam Leff <adam@leff.co>

* Remove bundler install during Appveyor tests (#217)

It appears that the Appveyor images changed recently and the Ruby
installs for all version (2.2, 2.3, and 2.4) already include Bundler
installed. Attempting to install it will hang the tests at an
interactive prompt asking if you want to override the install, and
--force'ing the install causes path issues and Rubygems won't be able
to activate the gem.

Removing the bundler install appears to solve the problem nicely.

Signed-off-by: Adam Leff <adam@leff.co>