Adds cached_client method in BaseConnection
Adds a helper method in base connection so you no longer have to
duplicate caching logic in any plugins you may create. Example:

```Ruby
def demo_client
  cache_client(:api_call, :demo_connection) do
    DemoClient.new(options)
  end
end
```

This will automatically check if the cache is enabled and return the the
cached object if it exists. If it doesn't exist then the cache will be
hydrated.

Signed-off-by: David McCown <dmccown@chef.io>