Fix Cisco IOS detection when banners lack a `\r\n` (#372)
This modifies some detection RegEx to not require `Cisco IOS Software`
to be at the start of a line.

During detection, the first command ran will contain the MOTD banner. So if the
banner was created without a newline before the delimiter character like so:

```
IOS_12_HARDENED(config)#banner motd |
Enter TEXT message.  End with the character '|'
A MOTD Banner|
IOS_12_HARDENED(config)#
```

It will show up in the output of `show version` like so:

```
A MOTD BannerCisco IOS Software...
```

If someone were to create a banner like this:

```
IOS_12_HARDENED(config)#banner motd |
Enter TEXT message.  End with the character '|'
A MOTD Banner
|
IOS_12_HARDENED(config)#
```

Then the detection RegEx would work because it would look like the
following in the output:

```
A MOTD Banner\r\nCisco IOS Software...
```

This modifies the detection to support both as both are valid.

Signed-off-by: Jerry Aldrich <jerryaldrichiii@gmail.com>