Fixes enable password and catches failure
On my switch, '\r\n' is being interpreted as two enter key presses, so
the enable password is being sent to the shell prompt on the switch, not
to the password prompt.

I'm not sure if it is different behavior on some routers or if this has
been broken for a while, given that inspec may not need enable. But
'show running-config' does require enable privileges.

Fixes #382

Signed-off-by: Bryan McLellan <btm@loftninjas.org>