Add data callbacks to WinRM and SSH
In order to support existing bootstrap functionality for `knife
bootstrap`s conversion to Train, clients need a method to receive
data from the target host as it is being generated.

This callback will be invoked for SSH when data arrives on the channel,
and for WinRM when data is returned on stdout.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>