Remove dependency on ed25519 gems
Having this here in train means that inspec requires it. That means that
an installation of InSpec into an existing Chef install requires
compilation tools on the node, which is not an option for many of our
users. We need to instead require this in chef-core and in the inspec
omnibus setup..

Signed-off-by: Tim Smith <tsmith@chef.io>