Fixed cisco_ios? returning false by default and other related broken tests.
Way back in the day, I put this method in in order for some stubbing
to work correctly. It probably shouldn't have gone into the PR tho
past getting the tests up and running.

What I figured out in this debugging pass was that there was some
infection from the tests caused by train using class-instance
variables that weren't getting reset when needed. Now I'm doing it on
every single test, which is probably overkill, but I'd rather have
stable and slow.

Signed-off-by: Ryan Davis <zenspider@chef.io>