Windows exit code should default to 0
Previously local Windows commands hardcoded the exit code as 0.
After changing this to get the real exit code; there are cases where
LASTEXITCODE is not set and is `$null`.

`$null` is probably the correct value to reflect in this case; but to
maintain compatibility with previous versions of train that returned 0,
this commit has train return 0 when LASTEXITCODE is `$null`

Signed-off-by: James Stocks <jstocks@chef.io>