Updates dependabot configs
Many of our dependencies should require a simple human check and then merge.

However there are some dependencies that we still want to be a manual process as they are quite tricky version to version.

Doing so through Github does not allow us to document these changes, so we are using dependabot configs to allow discussion and documentation of what may be ignored.

This was sparked by an ActiveSupport bump which would have broken ruby 2.4 support.

Signed-off-by: Nick Schwaderer <nschwaderer@chef.io>