Ensure file --> exist works with Solaris10
__Description__

See below issue for full context here. The short summary is that
Solaris10 does not support `test -e` flag in `sh`. This is a large
issue because 1. we support Solaris10 and 2. this is a silent failure.

The goal of this PR is to recognize a Solaris instance at the breaking
point, and then run the command with the provided `/usr/bin/test`
utility.

__Related Issues__

Fixes #587

__Run on my machine__

@lhasadreams provided fantastic guidance on the referenced issue for
replicating on your machine. In vagrant run the above control with
the vagrant box `tnarik/solaris10-minimal`:

```ruby
control "find_file_test_sol10-1" do
  impact 1.0
  describe command('test -e /etc/passwd') do
     its('exit_status') { should eq 0 }
  end
  describe file('/etc/passwd') do
    it{ should exist }
  end
end
```

(Note: make sure `sh` is your default shell in the `etc/passwd`)

On master both should fail. On my change second should pass.

Signed-off-by: Nick Schwaderer <nschwaderer@chef.io>