Speed up requires in non-omnibus Ruby installs
require is quite slow in Ruby and doing requires for things you've already required is also slow. We've used this simple hack in Chef to speed up our requires. In the omnibus installs we patch how rubygems works to make this somewhat pointless, but this will help non-omnibus installs of InSpec.

Signed-off-by: Tim Smith <tsmith@chef.io>