Make :forward_agent a valid SSH option
When I run `knife bootstrap --forward-agent …` chef calls Train::options
to fetch the valid options for an SSH connection and rejects any options
that are not listed as valid for Train::Transport::SSH:

    https://github.com/chef/chef/blob/v16.8.2/lib/chef/knife/bootstrap/train_connector.rb#L269-L280

One of the options validated is `:forward_agent`.  Since it is not
listed as a valid option by the SSH transport it can never be set on
any connection.

Without this change Train::Transport::SSH connections for `knife
bootstap` will always include `-o ForwardAgent=no` regardless of the
user's choices.

Previously :forward_agent was not listed as a valid option for the SSH
transport.  Now it is included as a valid option.

This means that `knife bootstrap --forward-agent` will now generate an
SSH command containing `-o ForwardAgent=yes`.

The option is false by default in order to match current train behavior.

Signed-off-by: Eric Hodel <drbrain@segment7.net>